Metashape - Extract helper class to dedicated files
Help readability of the main processing class and extract helper classes
to their own files.
