+++ metashape/accuracy.py
@@ -1,0 +1,9 @@
+import Metashape
+
+
+class Accuracy:
+    CAMERA_LOCATION = Metashape.Vector([0.5, 0.5, 0.5])  # in meter
+    CAMERA_ROTATION = Metashape.Vector([5, 5, 5])  # in degrees
+    MARKERS = Metashape.Vector([0.03, 0.03, 0.03])  # in meter
+    MARKER_PROJECTION = 1.0  # in pixel
+    SCALEBAR = 0.03  # in meter
