+++ metashape/depth_map_quality.py
@@ -1,0 +1,4 @@
+class DepthMapQuality:
+    ULTRA = 1
+    HIGH = 2
+    MEDIUM = 4
