+++ metashape/filter.py
@@ -1,0 +1,6 @@
+class Filter:
+    RECONSTRUCTION_UNCERTAINTY = 10.0  # no units
+    REPROJECTION_ERROR = 0.3  # in pixels
+    PROJECTION_ACCURACY = 5.0  # in pixels
+    DEPTH_MAP_MINIMUM = 1  # Count number
+    DENSE_CLOUD_POINT_SPACING = 0.00025  # in meters
