+++ metashape/image_matching.py
@@ -1,0 +1,4 @@
+class ImageMatching:
+    HIGHEST = 0
+    HIGH = 1
+    MEDIUM = 2
